gradio_ui/aixbt_interface/requirements.txt
gradio>=4.0.0
redis>=5.0.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
websockets>=12.0